Document that plugin distribution will be suspended
The Crowd 2 integration plugin is using dependencies that are closed
source and are not licensed with an OSI approved open source license.

The Jenkins project governance document and the plugin hosting page
state that we only deliver plugins with open source licenses.

The specific license of concern is:

* https://www.atlassian.com/legal/software-license-agreement
  Atlassian customer license

The same license of concern is used for:

* http://www.atlassian.com/end-user-agreement/
  Atlassian 3.0 End User License Agreement

From the `/plugin/crowd2/wrapper/thirdPartyLicenses` page of a Jenkins
installation that includes the Crowd 2 plugin:

Dependencies using the Atlassian Customer Agreement:

* com.atlassian.crowd:crowd-integration-api:5.1.5
* com.atlassian.crowd:crowd-integration-client-rest:5.1.5
* com.atlassian.crowd:embedded-crowd-api:5.1.5
* com.atlassian.crowd:crowd-integration-client-common:5.1.5

Dependencies using the Atlassian 3.0 End User License Agreement

* com.atlassian.collectors:atlassian-collectors-util:1.1